numlines changed from int to size_t

+++ TP0/includes/common.h
@@ -5,7 +5,7 @@ void show_help();
 
 void show_version();
 
-int process_file(char* file, char ***lineas, int *numlines);
+int process_file(char* file, char ***lineas, size_t *numlines);
 
 int save_file(char* file, char **lineas, size_t numLines);
